Refactor metrics class to top level
