+++ gnd/src/main/java/com/google/android/gnd/ui/home/OnBottomSheetSlideBehavior.java
@@ -17,13 +17,10 @@
 package com.google.android.gnd.ui.home;
 
 import android.content.Context;
-import android.graphics.drawable.Drawable;
 import android.util.AttributeSet;
 import android.view.View;
-import androidx.annotation.Nullable;
 import androidx.coordinatorlayout.widget.CoordinatorLayout;
 import com.google.android.gnd.R;
-import com.google.android.gnd.ui.common.BottomSheetBehavior;
 
 public abstract class OnBottomSheetSlideBehavior<V extends View>
     extends CoordinatorLayout.Behavior<V> {
@@ -43,83 +40,5 @@ public abstract class OnBottomSheetSlideBehavior<V extends View>
   public boolean onDependentViewChanged(CoordinatorLayout parent, V child, View bottomSheet) {
     onSheetScrolled(parent, child, new HomeScreenMetrics(parent, bottomSheet));
     return false;
-  }
-
-  public static class HomeScreenMetrics {
-    private final CoordinatorLayout parent;
-    private final View bottomSheet;
-
-    public HomeScreenMetrics(CoordinatorLayout parent, View bottomSheet) {
-      this.parent = parent;
-      this.bottomSheet = bottomSheet;
-    }
-
-    public static float scale(
-        float value, float before1, float before2, float after1, float after2) {
-      if (before1 > before2) {
-        return scale(value, before2, before1, after1, after2);
-      }
-      float minBefore = before1;
-      float maxBefore = before2;
-      float rangeBefore = maxBefore - minBefore;
-      float minAfter = Math.min(after1, after2);
-      float maxAfter = Math.max(after1, after2);
-      float rangeAfter = maxAfter - minAfter;
-      float ratio = (value - minBefore) / rangeBefore;
-      if (ratio > 1) {
-        ratio = 1;
-      }
-      if (ratio < 0) {
-        ratio = 0;
-      }
-      if (after1 > after2) {
-        ratio = 1 - ratio;
-      }
-      return ratio * rangeAfter + minAfter;
-    }
-
-    public float getVisibleRatio() {
-      return 1 - ((float) bottomSheet.getTop()) / ((float) parent.getHeight());
-    }
-
-    public void showWithSheet(Drawable view, float hideThreshold, float showThreshold) {
-      view.setAlpha((int) scale(getVisibleRatio(), hideThreshold, showThreshold, 0.0f, 255.0f));
-    }
-
-    public void hideWithSheet(Drawable view, float showThreshold, float hideThreshold) {
-      view.setAlpha((int) scale(getVisibleRatio(), showThreshold, hideThreshold, 255.0f, 0.0f));
-    }
-
-    public void showWithSheet(@Nullable View view, float hideThreshold, float showThreshold) {
-      if (view != null) {
-        view.setAlpha(scale(getVisibleRatio(), hideThreshold, showThreshold, 0.0f, 1.0f));
-      }
-    }
-
-    public void hideWithSheet(@Nullable View view, float showThreshold, float hideThreshold) {
-      if (view != null) {
-        view.setAlpha(scale(getVisibleRatio(), showThreshold, hideThreshold, 1.0f, 0.0f));
-      }
-    }
-
-    public int getToolbarHeight() {
-      return parent.findViewById(R.id.toolbar).getHeight();
-    }
-
-    public int getVisibleHeight() {
-      return Math.max(parent.getHeight() - bottomSheet.getTop(), 0);
-    }
-
-    public int getTop() {
-      return bottomSheet.getTop();
-    }
-
-    public int getTabLayoutTop() {
-      return bottomSheet.getTop() + bottomSheet.getPaddingTop();
-    }
-
-    public int getPeekHeight() {
-      return BottomSheetBehavior.from(bottomSheet).getPeekHeight();
-    }
   }
 }
