+++ gnd/src/main/java/com/google/android/gnd/ui/home/mapcontainer/MapContainerLayoutBehavior.java
@@ -22,6 +22,7 @@ import android.view.View;
 import android.widget.FrameLayout;
 import androidx.coordinatorlayout.widget.CoordinatorLayout;
 import com.google.android.gnd.R;
+import com.google.android.gnd.ui.home.HomeScreenMetrics;
 import com.google.android.gnd.ui.home.OnBottomSheetSlideBehavior;
 
 public class MapContainerLayoutBehavior extends OnBottomSheetSlideBehavior<FrameLayout> {
